\name{bigmelon-package}
\alias{bigmelon-package}
\alias{bigmelon}
\docType{package}
\title{
Write large-scale Illumina array datasets to CoreArray Genomic Data
Structure (GDS) data files for efficient storage, access, and modification.
}
\description{
Functions for storing Illumina array data as CoreArray Genomic Data Structure
(GDS) data files (via the gdsfmt package), appending these files , and
applying array normalization methods from the wateRmelon package.
}
\details{
\tabular{ll}{
Package: \tab bigmelon\cr
Type: \tab Package\cr
Version: \tab 1.1.13\cr
Date: \tab 2017-04-06\cr
License: \tab GPL3\cr
}
}
\author{
Tyler Gorrie-Stone Leonard C Schalkwyk, Ayden Saffari, Karim Malki.
Who to contact: <tgorri@essex.ac.uk>, <leonard.schalkwyk@kcl.ac.uk>

}
\references{
[1]Pidsley R, Wong CCY, Volta M, Lunnon K, Mill J, Schalkwyk LC:
A data-driven approach to preprocessing Illumina 450K methylation array data.
BMC genomics, 14(1), 293.
}

\seealso{
\code{\link{es2gds}}, \code{\link{dasen}}, \code{\link{wateRmelon}},
\code{\link{gdsfmt}}.,
\code{\link[methylumi]{methylumi}}.

}


\keyword{package}
