% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inout.R
\name{bigPepo}
\alias{bigPepo}
\title{tea.gds - description and validation for gds objects}
\usage{
bigPepo(path, gds, manifest, chunksize = NULL, force = TRUE, ...)
}
\description{
Checks the validity, file location and read/write status of a gds object
}
\details{
For DNAm use
with the bigmelon package.  For interactive use the default verbose 
option also prints
further details. Returns a list.
}
