\name{cantaloupe}
\alias{cantaloupe}
\alias{honeydew}
\docType{data}
\title{
Small MethyLumi 450k data sets for testing
}
\description{
Small MethyLumi 450k data sets intended for testing purposes only.
}
\usage{
data(cantaloupe)
data(honeydew)
}
\value{
Loads data into R
}
\format{
cantaloupe: MethyLumiSet with assayData containing 841 features, 3 samples.
honeydew: MethyLumiSet with assayData containing 841 features, 4 samples.
}

%\references{
%}
%\examples{
%}
\keyword{datasets}

