% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainsaw.R
\name{chainsaw}
\alias{chainsaw}
\title{Chainsaw -- modify gds file by subsetting all nodes
candidate replacement for `[.gds.class`}
\usage{
chainsaw(gfile, i, j, v = TRUE)
}
\description{
Chainsaw -- modify gds file by subsetting all nodes
candidate replacement for `[.gds.class`
}
