% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainsaw.R
\name{dim.gds.class}
\alias{dim.gds.class}
\title{dim.gds.class  S3 method returning dimensions of data represented by a gds file handle.}
\usage{
\method{dim}{gds.class}(gfile, v = FALSE)
}
\description{
dim.gds.class  S3 method returning dimensions of data represented by a gds file handle.
}
